backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pydantic>=1.8.0
python-dotenv>=0.19.0
PyPDF2>=2.0.0
python-docx>=0.8.11
chromadb>=0.4.0
openai>=1.0.0
sentence-transformers>=2.2.2
pydantic-settings>=2.0.0
google-generativeai>=0.3.0